+++ R/predSummary.R
@@ -3,16 +3,20 @@
 #' 
 #' Quick function to calculate performance metrics: confusion matrix, accuracy
 #' and balanced accuracy for classification; ROC AUC for binary classification;
-#' RMSE and R^2 for regression.
+#' RMSE and R^2 for regression. Multi-class AUC is returned for multinomial
+#' classification.
 #' 
 #' @param output data.frame with columns `testy` containing observed response
 #'   from test folds; `predy` predicted response; `predyp` (optional) predicted
 #'   probabilities for classification to calculate ROC AUC
 #' @return An object of class 'predSummary'. For classification a list is
 #'   returned containing the confusion matrix table and a vector containing
-#'   accuracy and balanced accuracy for classification, ROC AUC for binary
+#'   accuracy and balanced accuracy for classification, ROC AUC for 
 #'   classification. For regression a vector containing RMSE and R^2 is
 #'   returned.
+#' @details
+#' For multinomial classification, multi-class AUC as defined by Hand and Till
+#' is calculated using [pROC::multiclass.roc()].
 #' 
 #' @export
 predSummary <- function(output) {
@@ -33,7 +37,11 @@ predSummary <- function(output) {
       auc <- outputroc$auc
       metrics <- setNames(c(auc, acc, b_acc), c("AUC", "Accuracy", "Balanced accuracy"))
     } else {
-      metrics <- setNames(c(acc, b_acc), c("Accuracy", "Balanced accuracy"))
+      auc <- try(pROC::multiclass.roc(output$testy, output[, -c(1,2)])$auc,
+                 silent = TRUE)
+      if (inherits(auc, "try-error")) auc <- NA
+      metrics <- setNames(c(auc, acc, b_acc), c("Multiclass AUC", "Accuracy", 
+                                                "Balanced accuracy"))
     }
     summary <- list(table = cm, metrics = metrics)
   } else {